fix voucher template amount display

+++ model/voucher.go
@@ -12,9 +12,9 @@ import (
 	"gorm.io/gorm/clause"
 )
 
-func AmountReplace(original string, amount int64) string {
+func AmountReplace(original string, amount float64) string {
 	return util.TextReplace(original, map[string]string{
-		"amount": fmt.Sprintf("%d", amount),
+		"amount": fmt.Sprintf("%.2f", amount),
 	})
 }
